Revert "Implement redo feature (#175)"
This reverts commit 0b4970c595bcf096d962aaf9172950e7f74bb515.
