+++ src/main/java/seedu/address/logic/commands/undocommand/UndoCommand.java
@@ -1,5 +1,6 @@
-package seedu.address.logic.commands.historycommand;
+package seedu.address.logic.commands.undocommand;
 
+import seedu.address.commons.exceptions.AlfredModelHistoryException;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
@@ -9,12 +10,10 @@ import seedu.address.model.entity.CommandType;
 /**
  * Command that undoes the effects of the previous command, returning the model to its previous state.
  */
-public class HistoryCommand extends Command {
-    public static final String COMMAND_WORD = "history";
-    public static final String MESSAGE_SUCCESS = "History shown";
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Shows the history of previously-executed commands"
-                                                            + "that are undo-able and redo-able";
-    public static final String MESSAGE_SHOW_HISTORY = "Showing the history of your commands.";
+public class UndoCommand extends Command {
+    public static final String COMMAND_WORD = "undo";
+    public static final String MESSAGE_SUCCESS = "Undid 1 command";
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Undoes the previous command";
 
     /**
      * Executes the command and returns a CommandResult with a message.
@@ -23,6 +22,11 @@ public class HistoryCommand extends Command {
      * @throws CommandException
      */
     public CommandResult execute(Model model) throws CommandException {
-        return new CommandResult(MESSAGE_SHOW_HISTORY, CommandType.H);
+        try {
+            model.undo();
+            return new CommandResult(String.format(MESSAGE_SUCCESS), CommandType.H);
+        } catch (AlfredModelHistoryException e) {
+            throw new CommandException(e.getMessage());
+        }
     }
 }
